refactor(auth): moved validate-jwt to feature folder and centralized integration into server index
