+++ server/src/features/auth/index.ts
@@ -2,5 +2,5 @@ export * from "./controllers";
 export * from "./schemas";
 export * from "./services";
 export * from "./utils";
-export * from "./auth.middleware";
+export * from "./middleware";
 export * from "./auth.routes";
