+++ server/src/features/auth/middleware/index.ts
@@ -1,0 +1,2 @@
+export * from "./user-data-service.middleware";
+export * from "./validate-jwt";
