+++ server/src/features/auth/middleware/validate-jwt.ts
