+++ server/src/features/projects/project-tags.router.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { API, AUTH } from "@data";
-import { validateJWT, verifyRoles } from "@middleware";
+import { verifyRoles } from "@middleware";
 import { attachTagsCacheHandler } from "./projects.middleware";
 
 const PROJECT_TAG_ROUTES = API.PROJECT_TAG_ROUTES;
@@ -8,7 +8,6 @@ const { Guest, Student, Moderator, Administrator } = AUTH.ROLES_MAP;
 
 export const projectTagsRouter = express.Router();
 
-projectTagsRouter.use(validateJWT);
 projectTagsRouter.use(attachTagsCacheHandler);
 
 projectTagsRouter.get(
