+++ server/src/features/projects/projects.router.ts
@@ -1,6 +1,6 @@
 import express, { Request } from "express";
 import { API, AUTH } from "@data";
-import { validateJWT, verifyRoles } from "@middleware";
+import { verifyRoles } from "@middleware";
 import { attachProjectCachePageService } from "./projects.middleware";
 import { IProjectFilter } from "./services";
 import { IProjectDetails } from "./types";
@@ -10,7 +10,6 @@ const { Guest, Student, Moderator, Administrator } = AUTH.ROLES_MAP;
 
 export const projectsRouter = express.Router();
 
-projectsRouter.use(validateJWT);
 projectsRouter.use(attachProjectCachePageService);
 
 //  for testing
