+++ server/src/middleware/auth/index.ts
@@ -1,2 +1,1 @@
-export * from "./validate-jwt";
 export * from "./verify-roles";
